refactor: reorganize service/usage folder by separating different components
